+++ src/pages/helpers/HelpersPage.tsx
@@ -1,5 +1,8 @@
+/* eslint-disable prettier/prettier */
 import Checkbox from '@mui/material/Checkbox';
 import FormControlLabel from '@mui/material/FormControlLabel';
+import MenuItem from '@mui/material/MenuItem';
+import Select, {SelectChangeEvent} from '@mui/material/Select';
 import Stack from '@mui/material/Stack';
 import TextField from '@mui/material/TextField';
 import Typography from '@mui/material/Typography';
@@ -8,20 +11,29 @@ import React, {useContext, useState} from 'react';
 import SpacedTypography from '@app/components/SpacedTypography';
 import AuthenticationContext from '@app/context/AuthenticationContext';
 import useDelay from '@app/hooks/useDelay';
+import {getCurrentYear} from '@app/utils/date-utils';
 import {SEARCH_DELAY_MS} from '@app/utils/search-utils';
 
 import HelperTasksDataGrid from './HelperTasksDataGrid';
-import PageTitleWithNewTaskButton from './PageTitleWithTaskActions';
+import PageTitleWithNewAndCloneTaskButtons from './PageTitleWithTaskActions';
 
 const HelpersPage = () => {
   const currentUser = useContext(AuthenticationContext).currentUser;
+  const firstHelperAppYear = 2023;
+  const currentYear = getCurrentYear();
 
+  const [year, setYear] = useState<number | null>(currentYear);
   const [search, setSearch] = useState<string>('');
   const [showOnlyUpcoming, setShowOnlyUpcoming] = useState(true);
   const [showOnlyContactOrSignedUp, setShowOnlyContactOrSignedUp] =
     useState(false);
   const [showOnlyAvailable, setShowOnlyAvailable] = useState(false);
   const [showOnlyUnpublished, setShowOnlyUnpublished] = useState(false);
+
+  const onYearChange = (event: SelectChangeEvent) => {
+    setYear(event.target.value === 'ALL' ? null : parseInt(event.target.value));
+    setShowOnlyUpcoming(false);
+  };
 
   const onSearch = useDelay(
     SEARCH_DELAY_MS,
@@ -38,17 +50,38 @@ const HelpersPage = () => {
     handler(checked);
   };
 
+  const years = Array.from(
+    {length: currentYear - firstHelperAppYear + 1},
+    (_, i) => firstHelperAppYear + i
+  );
+
   return (
     <>
-      <PageTitleWithNewTaskButton value="Helper Tasks" />
+      <PageTitleWithNewAndCloneTaskButtons value="Helper Tasks" />
 
-      <SpacedTypography>
-        On this page you can sign up for surveillance and maintenance tasks.
-        Captain means Q-boat driver or the person who is organising the
-        execution of the task.
-      </SpacedTypography>
-
+      {/* TODO make the layout mobile friendly */}
       <Stack direction="row" alignItems="center" spacing={1} mt={2} mb={2}>
+        {currentUser.helpersAppAdminOrEditor && (
+          <>
+            <SpacedTypography>Year:</SpacedTypography>
+            <Select
+              defaultValue={currentYear}
+              onChange={onYearChange}
+              variant="outlined"
+              size="small"
+            >
+              {years.map(year => (
+                <MenuItem key={year} value={year}>
+                  {year}
+                </MenuItem>
+              ))}
+              <MenuItem key={9999} value="ALL">
+                ALL
+              </MenuItem>
+            </Select>
+          </>
+        )}
+
         <Typography>Search:</Typography>
         <TextField
           onChange={onSearch}
@@ -56,7 +89,8 @@ const HelpersPage = () => {
           label="Category, person, text..."
           size="small"
           sx={{
-            width: 230,
+            height: 40,
+            width: 200,
           }}
           className="ycc-helpers-search-input"
         />
@@ -69,6 +103,7 @@ const HelpersPage = () => {
                   setShowOnlyUpcoming(checked)
                 )
               }
+              size="small"
             />
           }
           label="Only upcoming"
@@ -85,6 +120,7 @@ const HelpersPage = () => {
                   }
                 })
               }
+              size="small"
             />
           }
           label="Only mine"
@@ -101,6 +137,7 @@ const HelpersPage = () => {
                   }
                 })
               }
+              size="small"
             />
           }
           label="Only available"
@@ -115,6 +152,7 @@ const HelpersPage = () => {
                     setShowOnlyUnpublished(checked)
                   )
                 }
+                size="small"
               />
             }
             label="Only unpublished"
@@ -123,6 +161,7 @@ const HelpersPage = () => {
       </Stack>
 
       <HelperTasksDataGrid
+        year={year}
         search={search}
         showOnlyUpcoming={showOnlyUpcoming}
         showOnlyContactOrSignedUp={showOnlyContactOrSignedUp}